+++ internal/appwindow/bridge_test.go
@@ -1,64 +1,12 @@
-package main
+package appwindow
 
 import (
-	"context"
-	"errors"
-	"fmt"
-	"os"
-	"os/exec"
-	"runtime"
-	"sync"
 	"testing"
-	"time"
 
-	"github.com/luxury-yacht/app/backend"
 	"github.com/luxury-yacht/app/internal/panelwindow"
-	"github.com/luxury-yacht/app/internal/sentry"
 	"github.com/stretchr/testify/require"
 	"github.com/wailsapp/wails/v3/pkg/application"
 )
-
-var (
-	testCompositionOnce sync.Once
-	testComposition     *applicationComposition
-)
-
-func sharedTestComposition() *applicationComposition {
-	testCompositionOnce.Do(func() {
-		testComposition = newApplicationComposition(&mainRecordingReporter{}, compositionOptions{
-			SingleInstance:         true,
-			SingleInstanceUniqueID: testSingleInstanceID(),
-		})
-	})
-	return testComposition
-}
-
-func testSingleInstanceID() string {
-	return fmt.Sprintf("%s.test.%d", applicationProductIdentifier, os.Getpid())
-}
-
-func TestNativeApplicationMenuIsInstalledOnlyOnDarwin(t *testing.T) {
-	for _, test := range []struct {
-		goos      string
-		wantCalls int
-	}{
-		{goos: "darwin", wantCalls: 1},
-		{goos: "windows"},
-		{goos: "linux"},
-	} {
-		t.Run(test.goos, func(t *testing.T) {
-			calls := 0
-			installNativeApplicationMenuForPlatform(test.goos, func() { calls++ })
-
-			require.Equal(t, test.wantCalls, calls)
-		})
-	}
-}
-
-type mainRecordingReporter struct {
-	panics     []any
-	exceptions []error
-}
 
 type recordingNativeWindowRegistry struct {
 	panelwindow.SharedWorkspaceCommands
@@ -179,56 +127,11 @@ func (registry *recordingNativeWindowRegistry) AcknowledgeApplicationQuitPreflig
 	return nil
 }
 
-func (*mainRecordingReporter) Enabled() bool                                   { return true }
-func (*mainRecordingReporter) SetEnabled(bool) error                           { return nil }
-func (*mainRecordingReporter) CaptureLogError(string, sentryreporting.Context) {}
-func (*mainRecordingReporter) AddBreadcrumb(sentryreporting.Breadcrumb)        {}
-func (*mainRecordingReporter) Shutdown(time.Duration) bool                     { return true }
-
-func (r *mainRecordingReporter) CaptureException(err error, _ sentryreporting.Context) {
-	r.exceptions = append(r.exceptions, err)
-}
-
-func (r *mainRecordingReporter) CapturePanic(recovered any, _ sentryreporting.Context) {
-	r.panics = append(r.panics, recovered)
-}
-
-func TestReportPanicCapturesAndRethrows(t *testing.T) {
-	reporter := &mainRecordingReporter{}
-
-	func() {
-		defer func() {
-			require.Equal(t, "boom", recover())
-		}()
-		func() {
-			defer reportPanic(reporter)
-			panic("boom")
-		}()
-	}()
-
-	require.Equal(t, []any{"boom"}, reporter.panics)
-}
-
-func TestReportRunErrorCapturesOnlyFailures(t *testing.T) {
-	reporter := &mainRecordingReporter{}
-	failure := errors.New("webview failed")
-
-	reportRunError(reporter, nil)
-	reportRunError(reporter, failure)
-
-	require.Equal(t, []error{failure}, reporter.exceptions)
-}
-
-func TestDefaultSentryReleaseUsesVersionedBuildIdentity(t *testing.T) {
-	require.Equal(t, "luxury-yacht@v1.2.3", defaultSentryRelease(" v1.2.3 "))
-	require.Empty(t, defaultSentryRelease("dev"))
-}
-
 func TestWindowRegistryBridgePreservesUnboundStartupSemantics(t *testing.T) {
-	bridge := &windowRegistryBridge{}
-	options := bridge.runtimeOptions(&mainRecordingReporter{}, backend.ApplicationUpdateOptions{})
+	bridge := &Bridge{}
+	options := bridge
 
-	require.True(t, bridge.prepareApplicationQuit())
+	require.True(t, bridge.PrepareApplicationQuit())
 	require.False(t, options.IsWorkspaceWindow("workspace-1"))
 	require.NotPanics(t, options.CreateWorkspaceWindow)
 	_, err := options.NativeWindowDescriptor("workspace-1")
@@ -265,13 +168,13 @@ func TestWindowRegistryBridgePreservesUnboundStartupSemantics(t *testing.T) {
 
 func TestWindowRegistryBridgeForwardsEveryRuntimeOperationAfterBinding(t *testing.T) {
 	registry := &recordingNativeWindowRegistry{}
-	bridge := &windowRegistryBridge{}
-	bridge.bind(registry)
-	options := bridge.runtimeOptions(&mainRecordingReporter{}, backend.ApplicationUpdateOptions{})
+	bridge := &Bridge{}
+	bridge.Bind(registry)
+	options := bridge
 	snapshot := panelwindow.GroupSnapshot{}
 
-	require.False(t, bridge.prepareApplicationQuit())
-	bridge.onSecondInstanceLaunch(application.SecondInstanceData{})
+	require.False(t, bridge.PrepareApplicationQuit())
+	bridge.OnSecondInstanceLaunch(application.SecondInstanceData{})
 	options.CreateWorkspaceWindow()
 	require.True(t, options.IsWorkspaceWindow("workspace-1"))
 	_, err := options.NativeWindowDescriptor("workspace-1")
@@ -314,101 +217,4 @@ func TestWindowRegistryBridgeForwardsEveryRuntimeOperationAfterBinding(t *testin
 		"fail-tab-transfer",
 		"acknowledge-quit",
 	}, registry.calls)
-}
-
-func TestNewSentryReporterStaysDisabledWhenBuildDisablesReporting(t *testing.T) {
-	t.Setenv("SENTRY_BACKEND_DSN", "https://runtime@example.com/2")
-
-	reporter, err := newSentryReporter(false, "https://embedded@example.com/1", "v1.2.3")
-
-	require.NoError(t, err)
-	require.False(t, reporter.Enabled())
-}
-
-func TestNewSentryReporterStartsDisabledUntilPersistedPreferenceLoads(t *testing.T) {
-	reporter, err := newSentryReporter(true, "https://embedded@example.com/1", "v1.2.3")
-
-	require.NoError(t, err)
-	require.False(t, reporter.Enabled())
-}
-
-func TestApplicationCompositionOwnsPeerWindowRegistryMenuAndService(t *testing.T) {
-	composition := sharedTestComposition()
-
-	require.NotNil(t, composition.application)
-	require.NotNil(t, composition.backend)
-	require.NotNil(t, composition.service)
-	require.NotNil(t, composition.operations)
-	require.NotNil(t, composition.menu)
-	if runtime.GOOS == "darwin" {
-		require.Equal(t, composition.menu, composition.application.Menu.GetApplicationMenu())
-	} else {
-		require.Nil(t, composition.application.Menu.GetApplicationMenu())
-	}
-
-	window, ok := composition.application.Window.GetByName("workspace-1")
-	require.True(t, ok)
-	require.NotNil(t, composition.windows)
-	require.Equal(t, "workspace-1", window.Name())
-	require.Equal(t, 1, composition.windows.Count())
-	config := composition.application.Config()
-	require.Len(t, config.Services, 1)
-	require.NotNil(t, config.Assets.Handler)
-	require.NotNil(t, config.ShouldQuit)
-	require.NotNil(t, config.SingleInstance)
-	require.Equal(t, testSingleInstanceID(), config.SingleInstance.UniqueID)
-	require.NotNil(t, config.SingleInstance.OnSecondInstanceLaunch)
-}
-
-func TestSingleInstanceUniqueIDDefaultsToProductIdentifier(t *testing.T) {
-	require.Equal(t, applicationProductIdentifier, singleInstanceUniqueID(""))
-	require.Equal(t, applicationProductIdentifier, singleInstanceUniqueID(" \t"))
-	require.Equal(t, "test-instance", singleInstanceUniqueID(" test-instance "))
-}
-
-type startupFailureProbeService struct {
-	name               string
-	startupErr         error
-	context            context.Context
-	shutdownContextErr error
-	sequence           *[]string
-}
-
-func (s *startupFailureProbeService) ServiceName() string { return s.name }
-
-func (s *startupFailureProbeService) ServiceStartup(ctx context.Context, _ application.ServiceOptions) error {
-	s.context = ctx
-	*s.sequence = append(*s.sequence, "start:"+s.name)
-	return s.startupErr
-}
-
-func (s *startupFailureProbeService) ServiceShutdown() error {
-	s.shutdownContextErr = s.context.Err()
-	*s.sequence = append(*s.sequence, "stop:"+s.name)
-	return nil
-}
-
-func TestApplicationRunRollsBackStartedServicesAfterStartupFailure(t *testing.T) {
-	const helperEnv = "LUXURY_YACHT_TEST_STARTUP_FAILURE"
-	if os.Getenv(helperEnv) != "1" {
-		command := exec.Command(os.Args[0], "-test.run=^TestApplicationRunRollsBackStartedServicesAfterStartupFailure$")
-		command.Env = append(os.Environ(), helperEnv+"=1")
-		output, err := command.CombinedOutput()
-		require.NoError(t, err, string(output))
-		return
-	}
-
-	sequence := []string{}
-	started := &startupFailureProbeService{name: "started", sequence: &sequence}
-	failure := errors.New("startup failed")
-	failing := &startupFailureProbeService{name: "failing", startupErr: failure, sequence: &sequence}
-	wailsApp := application.New(application.Options{ErrorHandler: func(error) {}})
-	wailsApp.RegisterService(application.NewService(started))
-	wailsApp.RegisterService(application.NewService(failing))
-
-	err := wailsApp.Run()
-
-	require.ErrorIs(t, err, failure)
-	require.Equal(t, []string{"start:started", "start:failing", "stop:started"}, sequence)
-	require.ErrorIs(t, started.shutdownContextErr, context.Canceled)
 }
