+++ internal/sentry/startup_test.go
@@ -1,0 +1,75 @@
+package sentryreporting
+
+import (
+	"errors"
+	"testing"
+	"time"
+
+	"github.com/stretchr/testify/require"
+)
+
+type mainRecordingReporter struct {
+	panics     []any
+	exceptions []error
+}
+
+func (*mainRecordingReporter) Enabled() bool                   { return true }
+func (*mainRecordingReporter) SetEnabled(bool) error           { return nil }
+func (*mainRecordingReporter) CaptureLogError(string, Context) {}
+func (*mainRecordingReporter) AddBreadcrumb(Breadcrumb)        {}
+func (*mainRecordingReporter) Shutdown(time.Duration) bool     { return true }
+
+func (r *mainRecordingReporter) CaptureException(err error, _ Context) {
+	r.exceptions = append(r.exceptions, err)
+}
+
+func (r *mainRecordingReporter) CapturePanic(recovered any, _ Context) {
+	r.panics = append(r.panics, recovered)
+}
+
+func TestReportPanicCapturesAndRethrows(t *testing.T) {
+	reporter := &mainRecordingReporter{}
+
+	func() {
+		defer func() {
+			require.Equal(t, "boom", recover())
+		}()
+		func() {
+			defer ReportPanic(reporter)
+			panic("boom")
+		}()
+	}()
+
+	require.Equal(t, []any{"boom"}, reporter.panics)
+}
+
+func TestReportRunErrorCapturesOnlyFailures(t *testing.T) {
+	reporter := &mainRecordingReporter{}
+	failure := errors.New("webview failed")
+
+	ReportRunError(reporter, nil)
+	ReportRunError(reporter, failure)
+
+	require.Equal(t, []error{failure}, reporter.exceptions)
+}
+
+func TestDefaultSentryReleaseUsesVersionedBuildIdentity(t *testing.T) {
+	require.Equal(t, "luxury-yacht@v1.2.3", defaultSentryRelease(" v1.2.3 "))
+	require.Empty(t, defaultSentryRelease("dev"))
+}
+
+func TestNewSentryReporterStaysDisabledWhenBuildDisablesReporting(t *testing.T) {
+	t.Setenv("SENTRY_BACKEND_DSN", "https://runtime@example.com/2")
+
+	reporter, err := NewStartupReporter(false, "https://embedded@example.com/1", "v1.2.3")
+
+	require.NoError(t, err)
+	require.False(t, reporter.Enabled())
+}
+
+func TestNewSentryReporterStartsDisabledUntilPersistedPreferenceLoads(t *testing.T) {
+	reporter, err := NewStartupReporter(true, "https://embedded@example.com/1", "v1.2.3")
+
+	require.NoError(t, err)
+	require.False(t, reporter.Enabled())
+}
